refactor: extract to a separate file
